+++ css/mix.css
@@ -2,7 +2,6 @@
 @import './modules/global.scss';
 @import './modules/animations.css';
 @import './modules/header.css';
-@import '../js/global/type/type.css';
 @import './modules/index.css';
 @import './modules/post.css';
 @import './modules/sidebar.css';
