+++ functions.php
@@ -114,6 +114,7 @@ function themeInit($widget)
     require_once 'libray/i18n/index.php';
     require_once 'libray/theme-helper.php';
     require_once 'libray/field.php';
+    require_once 'libray/short-code.php';
 }
