+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "typecho-theme-sagiri",
-  "version": "1.1.3",
+  "version": "1.1.4",
   "description": "Lovely theme for Typecho.",
   "main": "index.js",
   "directories": {
